requirements.txt
bottle
paste
PyYAML
requests
requests_toolbelt
